+++ dwtools/GaussianMixture.h
@@ -2,7 +2,7 @@
 #define _GaussianMixture_h_
 /* GaussianMixture.h
  *
- * Copyright (C) 2010-2011, 2015-2016 David Weenink
+ * Copyright (C) 2010-2017 David Weenink
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -34,7 +34,7 @@
 	Constraints for a Gaussian mixture:
 	 all covariances have the same 'dimension' parameter
 */
-autoGaussianMixture GaussianMixture_create (long numberOfComponents, long dimension, long storage);
+autoGaussianMixture GaussianMixture_create (integer numberOfComponents, integer dimension, integer storage);
 /* Start each function with expand and end with unExpand */
 
 void GaussianMixture_expandPCA (GaussianMixture me);
@@ -42,18 +42,18 @@ void GaussianMixture_expandPCA (GaussianMixture me);
 void GaussianMixture_unExpandPCA (GaussianMixture me);
 
 void GaussianMixture_drawConcentrationEllipses (GaussianMixture me, Graphics g,
-	double scale, int confidence, char32 *label, int pcaDirections, long d1, long d2,
+	double scale, int confidence, char32 *label, int pcaDirections, integer d1, integer d2,
 	double xmin, double xmax, double ymin, double ymax, int fontSize, int garnish);
 
 void GaussianMixture_and_PCA_drawConcentrationEllipses (GaussianMixture me, PCA him, Graphics g,
-	double scale, int confidence, char32 *label, long d1, long d2,
+	double scale, int confidence, char32 *label, integer d1, integer d2,
 	double xmin, double xmax, double ymin, double ymax, int fontSize, int garnish);
 
-void GaussianMixture_drawMarginalPdf (GaussianMixture me, Graphics g, long d, double xmin, double xmax, double ymin, double ymax, long npoints, long nbins, int garnish);
+void GaussianMixture_drawMarginalPdf (GaussianMixture me, Graphics g, integer d, double xmin, double xmax, double ymin, double ymax, integer npoints, integer nbins, int garnish);
 
-void GaussianMixture_and_PCA_drawMarginalPdf (GaussianMixture me, PCA him, Graphics g, long d, double xmin, double xmax, double ymin, double ymax, long npoints, long nbins, int garnish);
+void GaussianMixture_and_PCA_drawMarginalPdf (GaussianMixture me, PCA him, Graphics g, integer d, double xmin, double xmax, double ymin, double ymax, integer npoints, integer nbins, int garnish);
 
-autoGaussianMixture TableOfReal_to_GaussianMixture_fromRowLabels (TableOfReal me, long storage);
+autoGaussianMixture TableOfReal_to_GaussianMixture_fromRowLabels (TableOfReal me, integer storage);
 
 void GaussianMixture_initialGuess (GaussianMixture me, TableOfReal thee, double nSigmas, double ru_range);
 /*
@@ -75,13 +75,13 @@ void GaussianMixture_initialGuess (GaussianMixture me, TableOfReal thee, double 
 
 const char32 *GaussianMixture_criterionText (int criterion);
 
-autoGaussianMixture TableOfReal_to_GaussianMixture (TableOfReal me, long numberOfComponents, double delta_lnp, long maxNumberOfIterations, double lambda, int storage, int criterion);
+autoGaussianMixture TableOfReal_to_GaussianMixture (TableOfReal me, integer numberOfComponents, double delta_lnp, integer maxNumberOfIterations, double lambda, int storage, int criterion);
 
-void GaussianMixture_and_TableOfReal_improveLikelihood (GaussianMixture me, TableOfReal thee, double delta_lnp, long maxNumberOfIterations, double lambda, int criterion);
+void GaussianMixture_and_TableOfReal_improveLikelihood (GaussianMixture me, TableOfReal thee, double delta_lnp, integer maxNumberOfIterations, double lambda, int criterion);
 
-autoGaussianMixture GaussianMixture_and_TableOfReal_to_GaussianMixture_CEMM (GaussianMixture me, TableOfReal thee, long minNumberOfComponents, double delta_l, long maxNumberOfIterations, double lambda, int criterion);
+autoGaussianMixture GaussianMixture_and_TableOfReal_to_GaussianMixture_CEMM (GaussianMixture me, TableOfReal thee, integer minNumberOfComponents, double delta_l, integer maxNumberOfIterations, double lambda, int criterion);
 
-void GaussianMixture_splitComponent (GaussianMixture me, long component);
+void GaussianMixture_splitComponent (GaussianMixture me, integer component);
 
 autoClassificationTable GaussianMixture_and_TableOfReal_to_ClassificationTable (GaussianMixture me, TableOfReal thee);
 
@@ -105,7 +105,7 @@ autoCovariance GaussianMixture_to_Covariance_between (GaussianMixture me);
 autoCovariance GaussianMixture_to_Covariance_within (GaussianMixture me);
 
 
-autoCovariance GaussianMixture_extractComponent(GaussianMixture me, long component);
+autoCovariance GaussianMixture_extractComponent(GaussianMixture me, integer component);
 
 autoTableOfReal GaussianMixture_extractCentroids (GaussianMixture me);
 
@@ -113,19 +113,19 @@ autoTableOfReal GaussianMixture_extractMixingProbabilities (GaussianMixture me);
 
 autoPCA GaussianMixture_to_PCA (GaussianMixture me);
 
-autoMatrix GaussianMixture_and_PCA_to_Matrix_density (GaussianMixture me, PCA pca, long d1, long d2, double xmin, double xmax, long nx, double ymin, double ymax, long ny);
+autoMatrix GaussianMixture_and_PCA_to_Matrix_density (GaussianMixture me, PCA pca, integer d1, integer d2, double xmin, double xmax, integer nx, double ymin, double ymax, integer ny);
 
-void GaussianMixture_and_PCA_getIntervalsAlongDirections (GaussianMixture me, PCA thee, long d1, long d2, double nsigmas, double *xmin, double *xmax, double *ymin, double *ymax);
+void GaussianMixture_and_PCA_getIntervalsAlongDirections (GaussianMixture me, PCA thee, integer d1, integer d2, double nsigmas, double *xmin, double *xmax, double *ymin, double *ymax);
 
-void GaussianMixture_and_PCA_getIntervalAlongDirection (GaussianMixture me, PCA thee, long d, double nsigmas, double *xmin, double *xmax);
+void GaussianMixture_and_PCA_getIntervalAlongDirection (GaussianMixture me, PCA thee, integer d, double nsigmas, double *xmin, double *xmax);
 
-void GaussianMixture_getIntervalAlongDirection (GaussianMixture me, long d, double nsigmas, double *xmin, double *xmax);
+void GaussianMixture_getIntervalAlongDirection (GaussianMixture me, integer d, double nsigmas, double *xmin, double *xmax);
 
-void GaussianMixture_getIntervalsAlongDirections (GaussianMixture me, long d1, long d2, double nsigmas, double *xmin, double *xmax, double *ymin, double *ymax);
+void GaussianMixture_getIntervalsAlongDirections (GaussianMixture me, integer d1, integer d2, double nsigmas, double *xmin, double *xmax, double *ymin, double *ymax);
 
 /* with on demand expand of pca ! */
 int GaussianMixture_generateOneVector (GaussianMixture me, double *c, char32 **covname, double *buf);
 
-autoTableOfReal GaussianMixture_to_TableOfReal_randomSampling (GaussianMixture me, long numberOfPoints);
+autoTableOfReal GaussianMixture_to_TableOfReal_randomSampling (GaussianMixture me, integer numberOfPoints);
 
 #endif /* _GaussianMixture_h_ */
